it works after refactoring!
